Move manager/interface and add module docstring
